+++ docs/examples/index.rst
@@ -5,4 +5,3 @@ Examples
    :maxdepth: 1
 
    maps
-   network_analysis_demo_template
