+++ docs/examples/py_to_md.py
@@ -53,12 +53,9 @@ def py_to_md(
     os.remove(nb_file)
 
 
-file = "network_analysis_examples"
-py_to_md(file, move_n_folders_up=0)
-ss
 file = "maps"
 py_to_md(file, move_n_folders_up=0)
-ss
+
 file = "network_analysis_examples"
 py_to_md(file, move_n_folders_up=0)
