+++ docs/reference/geopandas_tools/index.rst
@@ -1,11 +1,13 @@
 Geopandas tools
 ================
 
+This page contains the vector geometry operations in sgis. The exception is line
+geometry operations, which is located in the network analysis section.
+
 .. toctree::
    :maxdepth: 3
 
    general
-   line_operations
    polygon_operations
    point_operations
    neighbors
