--- docs/reference/geopandas_tools/line_operations.rst
@@ -1,7 +1,0 @@
-Line geometry operations
-========================
-
-.. automodule:: sgis.geopandas_tools.line_operations
-    :members:
-    :exclude-members:
-    :show-inheritance:
