+++ docs/reference/networkanalysis/cutting_lines.rst
@@ -1,0 +1,10 @@
+Cutting lines
+=============
+
+.. automodule:: sgis.networkanalysis.cutting_lines
+    :members:
+    :exclude-members:
+    :show-inheritance:
+    :noindex:
+
+.. Comment TODO: Investigate why noindex is needed, fix problem and remove
