+++ docs/reference/networkanalysis/finding_isolated_networks.rst
@@ -1,0 +1,10 @@
+Finding isolated networks
+==========================
+
+.. automodule:: sgis.networkanalysis.finding_isolated_networks
+    :members:
+    :exclude-members:
+    :show-inheritance:
+    :noindex:
+
+.. Comment TODO: Investigate why noindex is needed, fix problem and remove
