+++ docs/reference/networkanalysis/networkanalysis.rst
@@ -1,6 +1,8 @@
 NetworkAnalysis
 ===============
 
+Class for doing network analysis.
+
 .. autoclass:: sgis.networkanalysis.networkanalysis.NetworkAnalysis
     :members:
     :exclude-members:
