+++ docs/reference/networkanalysis/networkanalysisrules.rst
@@ -1,6 +1,8 @@
 NetworkAnalysisRules
 ====================
 
+Sets the rules for the NetworkAnalysis.
+
 .. autoclass:: sgis.networkanalysis.networkanalysisrules.NetworkAnalysisRules
     :members:
     :exclude-members:
