+++ docs/reference/networkanalysis/nodes.rst
@@ -1,0 +1,8 @@
+Nodes
+=====
+
+.. automodule:: sgis.networkanalysis.nodes
+    :members:
+    :exclude-members:
+    :show-inheritance:
+    :noindex:
